+++ src/containers/App/components/Home/Home.css
@@ -144,3 +144,11 @@
     border: none;
 }
 
+.JoinForm__Home {
+    width: 600px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    border-radius: 20px;
+    background-color: var(--amarillito);
+}
